requirements.txt
discord.py>=2.3.2
openai>=1.3.5
pandas>=2.2.2
python-dotenv>=1.0.1